Adding support for user in state
