+++ sia-react-native/App.js
@@ -1,29 +1,10 @@
-// import React from 'react';
-// import { StyleSheet, Text, View } from 'react-native';
-//
-// export default function App() {
-//   return (
-//     <View style={styles.container}>
-//       <Text>App is up and running!</Text>
-//     </View>
-//   );
-// }
-//
-// const styles = StyleSheet.create({
-//   container: {
-//     flex: 1,
-//     backgroundColor: '#fff',
-//     alignItems: 'center',
-//     justifyContent: 'center',
-//   },
-// });
 import React from 'react'
 import { createStore, applyMiddleware } from 'redux'
 import { Provider } from 'react-redux'
 import thunkMiddleware from 'redux-thunk'
 
 import SwitchNavigator from './navigation/SwitchNavigator'
-import reducer from './reducers'
+import reducer from './reducers/index'
 
 const middleware = applyMiddleware(thunkMiddleware)
 const store = createStore(reducer, middleware)
