+++ sia-react-native/reducers/auth.js
@@ -1,0 +1,17 @@
+import { combineReducers } from 'redux'
+import { LOGIN, SIGNUP, UPDATE_EMAIL, UPDATE_PASSWORD } from '../actions/auth'
+
+export default function auth(state = {}, action) {
+    switch (action.type) {
+        case LOGIN:
+            return action.payload
+        case SIGNUP:
+            return action.payload
+        case UPDATE_EMAIL:
+            return { ...state, email: action.payload }
+        case UPDATE_PASSWORD:
+            return { ...state, password: action.payload }
+        default:
+            return state
+    }
+}
