+++ sia-react-native/screens/Home.js
@@ -8,23 +8,31 @@ import {
   TouchableOpacity
 } from 'react-native'
 import {
+  bindActionCreators
+} from 'redux'
+import {
   connect
 } from 'react-redux'
 import Firebase from '../Firebase'
 import Backend from '../Backend.js'
 import user from '../classes/user.js'
-import Notification from '../notification/notification'
+import {
+  getUser
+} from '../actions/user'
 
 class Home extends React.Component {
   handleSignout = () => {
+    // this.props.getUser()
     Firebase.auth().signOut()
     this.props.navigation.navigate('Login')
   }
   render() {
     return ( <View style={styles.container}>
 				<Text>Profile Screen</Text>
-				<Text>{this.props.user.email}</Text>
-        <Text>{this.props.user.uid}</Text>
+				<Text>{this.props.auth.email}</Text>
+        <Text>{this.props.auth.uid}</Text>
+        <Text>{this.props.user.id}</Text>
+        <Text>{this.props.user.first_name}</Text>
 				<Button title='Logout' onPress={this.handleSignout} />
         <TouchableOpacity>
           <Notification></Notification>
@@ -43,10 +51,17 @@ const styles = StyleSheet.create({
   }
 })
 
+const mapDispatchToProps = dispatch => {
+  return bindActionCreators({
+    getUser
+  }, dispatch)
+}
+
 const mapStateToProps = state => {
   return {
+    auth: state.auth,
     user: state.user
   }
 }
 
-export default connect(mapStateToProps)(Home)
+export default connect(mapStateToProps, mapDispatchToProps)(Home)
