+++ sia-react-native/screens/Signup.js
@@ -19,6 +19,9 @@ import {
   updateEmail,
   updatePassword,
   signup
+} from '../actions/auth'
+import {
+  getUser
 } from '../actions/user'
 
 class Signup extends React.Component {
@@ -28,7 +31,9 @@ class Signup extends React.Component {
     password: ""
   }
   handleSignUp = () => {
-    this.props.signup()
+    this.props.signup().then( () => {
+      this.props.getUser()
+    })
     this.props.navigation.navigate('Home')
   }
 
@@ -36,14 +41,14 @@ class Signup extends React.Component {
     return ( <View style={styles.container}>
 				<TextInput
 					style={styles.inputBox}
-					value={this.props.user.email}
+					value={this.props.auth.email}
 					onChangeText={email => this.props.updateEmail(email)}
 					placeholder='Email'
 					autoCapitalize='none'
 				/>
 				<TextInput
 					style={styles.inputBox}
-					value={this.props.user.password}
+					value={this.props.auth.password}
 					onChangeText={password => this.props.updatePassword(password)}
 					placeholder='Password'
 					secureTextEntry={true}
@@ -97,12 +102,14 @@ const mapDispatchToProps = dispatch => {
   return bindActionCreators({
     updateEmail,
     updatePassword,
-    signup
+    signup,
+    getUser
   }, dispatch)
 }
 
 const mapStateToProps = state => {
   return {
+    auth: state.auth,
     user: state.user
   }
 }
